% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootmi.R
\name{bootmi_analyse}
\alias{bootmi_analyse}
\title{Analyse bootstrapped and imputed estimates}
\usage{
bootmi_analyse(imps, analysisfun, ...)
}
\arguments{
\item{imps}{The list of imputed datasets returned by \code{impute}}

\item{analysisfun}{A function which when applied to a single dataset returns
the estimate of the parameter(s) of interest.}

\item{...}{Other parameters that are to be passed through to \code{analysisfun}.}
}
\value{
A vector containing the point estimate(s), variance estimates, and
degrees of freedom.
}
\description{
Analyse bootstrapped and imputed estimates
}
