% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootSmcfcs.R
\name{bootSmcfcs}
\alias{bootSmcfcs}
\title{Bootstrap then impute using smcfcs}
\usage{
bootSmcfcs(obsdata, nBoot = 200, nImp = 2, ...)
}
\arguments{
\item{obsdata}{The data frame to be imputed.}

\item{nBoot}{The number of bootstrap samples to take. It is recommended
that you use a minimum of 200.}

\item{nImp}{The number of times to impute each bootstrap sample. Two
is recommended.}

\item{...}{Other arguments that are to be passed to \code{smcfcs}.}
}
\value{
A list of imputed datasets.
}
\description{
Bootstraps an incomplete dataset and then imputes each bootstrap a number
of times using the smcfcs package. The resulting list of bootstrapped then
imputed datasets can be analysed with \code{\link{bootImputeAnalyse}}.
To run this function requires the \code{smcfcs} package to be installed.
}
\examples{
if (requireNamespace("smcfcs", quietly = TRUE)) {
  library(smcfcs)
  set.seed(564764)
  head(ex_linquad)
  #bootstrap 10 times and impute each twice
  imps <- bootSmcfcs(ex_linquad, nBoot=10, nImp=2,
                     smtype="lm", smformula="y~z+x+xsq",
                     method=c("","","norm","x^2",""))
}
}
