% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootmi.R
\name{impute}
\alias{impute}
\title{Bootstrap then impute a partially observed dataset.}
\usage{
impute(obsdata, impfun, B = 200, M = 2, ...)
}
\arguments{
\item{obsdata}{The data frame to be imputed.}

\item{impfun}{A function which will will impute the missing values.}

\item{B}{The number of bootstrap samples to take.}

\item{M}{The number of times to impute each bootstrap sample.}

\item{...}{Other parameters that are to be passed through to \code{impfun}.}
}
\value{
A list of imputed datasets
}
\description{
Bootstrap then impute a partially observed dataset.
}
