% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootmi.R
\name{bootImpute}
\alias{bootImpute}
\title{Bootstrap then impute an incomplete dataset}
\usage{
bootImpute(obsdata, impfun, nBoot = 200, nImp = 2, ...)
}
\arguments{
\item{obsdata}{The data frame to be imputed.}

\item{impfun}{A function which when passed an incomplete dataset will
return a single imputed dataset.}

\item{...}{Other parameters that are to be passed through to \code{impfun}.}

\item{B}{The number of bootstrap samples to take.}

\item{M}{The number of times to impute each bootstrap sample.}
}
\value{
A list of imputed datasets
}
\description{
Bootstrap then impute an incomplete dataset
}
