% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootmi.R
\name{bootImputeAnalyse}
\alias{bootImputeAnalyse}
\title{Analyse bootstrapped and imputed estimates}
\usage{
bootImputeAnalyse(imps, analysisfun, ..., quiet = FALSE)
}
\arguments{
\item{imps}{The list of imputed datasets returned by \code{bootImpute}}

\item{analysisfun}{A function which when applied to a single dataset returns
the estimate of the parameter(s) of interest.}

\item{...}{Other parameters that are to be passed through to \code{analysisfun}.}

\item{quiet}{Specify whether to print output or not.}
}
\value{
A vector containing the point estimate(s), variance estimates, and
degrees of freedom.
}
\description{
Analyse bootstrapped and imputed estimates
}
